refactor: Lotto 클래스 변경, 메서드 수정, 파일명 수정
- public으로 관리하던 lottoNumberList -> private class filed로 변경
- createRandomNumberList -> set을 사용해서 리펙터링
- LootoApp -> LottoApp
