+++ src/js/LottoApp.js
